Qt style: Use an QImage that is backed by a SharedArray
This way we have one copy less of the data

+++ sixtyfps_runtime/corelib/graphics.rs
@@ -231,7 +231,7 @@ pub enum Resource {
     EmbeddedData(super::slice::Slice<'static, u8>),
     /// Raw ARGB
     #[allow(missing_docs)]
-    EmbeddedRgbaImage { width: u32, height: u32, data: super::sharedarray::SharedArray<u8> },
+    EmbeddedRgbaImage { width: u32, height: u32, data: super::sharedarray::SharedArray<u32> },
 }
 
 impl Default for Resource {
